pyproject.toml
[tool.poetry]
name = "resolve-ec2-id"
version = "1.0.0"
description = "Look up the EC2 instance ID given an instance name."
authors = ["Ben Mares <services-resolve-ec2-id@tensorial.com>"]
license = "MIT"
readme = "README.md"
include = ["py.typed"]

[tool.poetry.dependencies]
python = "^3.7"
typer = "^0.4.0"
typing-extensions = { version = "^4.0.1", python = "<3.8" }

[tool.poetry.dev-dependencies]
pre-commit = "^2.15.0"
tox = "^3.24.5"
tox-conda = "^0.9.1"
pytest = "^6.2.5"
black = "^21.12b0"
moto = {extras = ["ec2"], version = "^2.3.0"}
boto3-stubs = {extras = ["ec2"], version = "^1.20.26"}
pytest-cases = "^3.6.5"

[tool.poetry.scripts]
resolve-ec2-id = "resolve_ec2_id:app"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
